+++ SDL_project_0/RenderObject.cpp
@@ -1,11 +1,11 @@
 #include "stdafx.h"
 #include "RenderObject.h"
 
-RenderObject::RenderObject(int iD, std::string n, Transform t, Vector2i s, shared_Texture tx) :
+RenderObject::RenderObject(uint32_t iD, std::string n, Transform t, Vector2i s, shared_Texture tx) :
 	Object(iD, n, t),
 	screenSize(s),
 	texture(tx),
-	sortingPriority((int)ESortingPriority::AVERAGE),
+	sortingPriority((uint16_t)ESortingPriority::AVERAGE),
 	blendMode(SDL_BLENDMODE_BLEND)
 {
 	UpdateRect();
@@ -15,8 +15,8 @@ void RenderObject::UpdateRect() {
 	auto transf = GetUnifiedTransform();
 	rect.x = transf.position.x;
 	rect.y = transf.position.y;
-	rect.w = (int)(transf.scale.x * screenSize.x);
-	rect.h = (int)(transf.scale.y * screenSize.y);
+	rect.w = (uint32_t)(transf.scale.x * screenSize.x);
+	rect.h = (uint32_t)(transf.scale.y * screenSize.y);
 }
 
 
@@ -27,7 +27,7 @@ RenderObject& RenderObject::SetBlendMode(SDL_BlendMode mode) {
 	}
 	return *this;
 }
-RenderObject& RenderObject::SetSortingPriority(ESortingPriority priority, int offset ) {
-	sortingPriority = (int)priority + offset;
+RenderObject& RenderObject::SetSortingPriority(ESortingPriority priority, int_fast16_t offset ) {
+	sortingPriority = (int_fast16_t)priority + offset;
 	return *this;
 }
